Classi BL ed ID
-Cambiati tutti gli ID da int a long
-Select all su ClsPiattoBL e ClsStrumentoMusicaleBL

(questa commit andrebbe su strumentiMusicali, mi sono sbagliato ramo)

+++ NegozioStrumentiMusicali_Cappelloni-DiBernardo/DMO/ClsBatteria.cs
@@ -17,7 +17,7 @@ namespace NegozioStrumentiMusicali
 
         }
         //Costruttore ereditato dalla classe madre
-        public ClsBatteria(int id) : base(id)
+        public ClsBatteria(long id) : base(id)
         {
 
         }
