+++ NegozioStrumentiMusicali_Cappelloni-DiBernardo/DMO/ClsBatteriaPiatto.cs
@@ -9,16 +9,16 @@ namespace NegozioStrumentiMusicali
     public class ClsBatteriaPiatto
     {
         #region Attributi
-        private int _id;
-        private int _batteriaID;
-        private int _piattoID;
+        private long _id;
+        private long _batteriaID;
+        private long _piattoID;
 
         #endregion
 
         #region Proprietà
-        public int ID { get => _id; set => _id = value; }
-        public int BatteriaID { get => _batteriaID; set => _batteriaID = value; }
-        public int PiattoID { get => _piattoID; set => _piattoID = value; }
+        public long ID { get => _id; set => _id = value; }
+        public long BatteriaID { get => _batteriaID; set => _batteriaID = value; }
+        public long PiattoID { get => _piattoID; set => _piattoID = value; }
 
         #endregion
 
@@ -28,7 +28,7 @@ namespace NegozioStrumentiMusicali
 
         }
 
-        public ClsBatteriaPiatto(int id)
+        public ClsBatteriaPiatto(long id)
         {
             ID = id;
         }
